Full update with forumla and working functionality
From basic addition program, the program now accepts user input and
calulates the correct numbers, rounded to two decimal places.

The forumlas appear to be implemented correctly, although this needs
verifiying.

The button group to select payment schedule has been created, but the
functionality has not been tied into the model and action listener. This
needs to be completed. Currently, the payment schedule is still
hardcoded as a final double PAY_FREQ.