Add schedule window with scrollbar, check calcs
The calculations were checked with various sources, appear to correct
with a few potential rounding errors.

A view schedule button was added that triggers another view that will
show a table with payment schedule values. The table is set up, along
with a scrollbar, but the data is not.

To do:

 - set up data for table
 - look into seperate button panel for closing
 - clean up file structure, and make sure everything is placed according
   to the MVC model
 - double check calculations
 - testing