Add more error checking, and tool tip comments
Now, if negative or very low mortgage values are entered that result in
a irrelevant mortgage schedule, such as an ammortization period of 0 or
-1, an error box is shown and the schedule is not displayed. Without
this, console errors are generated and the window does not display
anyway.

In addition, tool tip comments were added to the text fields with
suggestions to increase usability.